feat: organizando os fragments em um novo package
