+++ app/src/main/java/com/educa/Quiz.kt
@@ -3,9 +3,6 @@ package com.educa
 import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
 
-
-
-
 class Quiz : AppCompatActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
