+++ app/src/main/java/com/educa/Reading.kt
@@ -5,7 +5,6 @@ import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
 import android.widget.Button
 import android.widget.TextView
-import org.w3c.dom.Text
 
 class Reading : AppCompatActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
