+++ app/src/main/java/com/educa/Redirect.kt
@@ -5,7 +5,6 @@ import android.net.Uri
 import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
 import android.widget.Button
-import android.widget.TextView
 
 class Redirect : AppCompatActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
