+++ app/src/main/java/com/educa/ui/fragments/FragmentButtonSeeTopics.kt
@@ -1,18 +1,23 @@
-package com.educa
+package com.educa.ui.fragments
 
 import android.os.Bundle
 import androidx.fragment.app.Fragment
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
+import com.educa.R
 
 // TODO: Rename parameter arguments, choose names that match
 // the fragment initialization parameters, e.g. ARG_ITEM_NUMBER
 private const val ARG_PARAM1 = "param1"
 private const val ARG_PARAM2 = "param2"
 
-
-class FragmentSalutation : Fragment() {
+/**
+ * A simple [Fragment] subclass.
+ * Use the [ButtonSeeTopics.newInstance] factory method to
+ * create an instance of this fragment.
+ */
+class ButtonSeeTopics : Fragment() {
     // TODO: Rename and change types of parameters
     private var param1: String? = null
     private var param2: String? = null
@@ -30,7 +35,7 @@ class FragmentSalutation : Fragment() {
         savedInstanceState: Bundle?
     ): View? {
         // Inflate the layout for this fragment
-        return inflater.inflate(R.layout.fragment_salutation, container, false)
+        return inflater.inflate(R.layout.fragment_button_see_topics, container, false)
     }
 
     companion object {
@@ -40,12 +45,12 @@ class FragmentSalutation : Fragment() {
          *
          * @param param1 Parameter 1.
          * @param param2 Parameter 2.
-         * @return A new instance of fragment SaudacaoFragment.
+         * @return A new instance of fragment ButtonSeeTopics.
          */
         // TODO: Rename and change types and number of parameters
         @JvmStatic
         fun newInstance(param1: String, param2: String) =
-            FragmentSalutation().apply {
+            ButtonSeeTopics().apply {
                 arguments = Bundle().apply {
                     putString(ARG_PARAM1, param1)
                     putString(ARG_PARAM2, param2)
