+++ app/src/main/java/com/educa/ui/fragments/FragmentEditContent.kt
@@ -1,4 +1,4 @@
-package com.educa
+package com.educa.ui.fragments
 
 import android.os.Bundle
 import androidx.fragment.app.Fragment
@@ -7,6 +7,7 @@ import android.view.View
 import android.view.ViewGroup
 import android.widget.ImageButton
 import androidx.appcompat.app.AppCompatActivity
+import com.educa.R
 
 class FragmentEditContent : Fragment() {
