+++ app/src/main/java/com/educa/ui/fragments/FragmentLikert.kt
@@ -1,10 +1,11 @@
-package com.educa
+package com.educa.ui.fragments
 
 import android.os.Bundle
 import androidx.fragment.app.Fragment
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
+import com.educa.R
 
 // TODO: Rename parameter arguments, choose names that match
 // the fragment initialization parameters, e.g. ARG_ITEM_NUMBER
@@ -13,10 +14,10 @@ private const val ARG_PARAM2 = "param2"
 
 /**
  * A simple [Fragment] subclass.
- * Use the [ButtonSeeTopics.newInstance] factory method to
+ * Use the [likert.newInstance] factory method to
  * create an instance of this fragment.
  */
-class ButtonSeeTopics : Fragment() {
+class likert : Fragment() {
     // TODO: Rename and change types of parameters
     private var param1: String? = null
     private var param2: String? = null
@@ -29,12 +30,13 @@ class ButtonSeeTopics : Fragment() {
         }
     }
 
+
     override fun onCreateView(
         inflater: LayoutInflater, container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View? {
         // Inflate the layout for this fragment
-        return inflater.inflate(R.layout.fragment_button_see_topics, container, false)
+        return inflater.inflate(R.layout.fragment_likert, container, false)
     }
 
     companion object {
@@ -44,12 +46,12 @@ class ButtonSeeTopics : Fragment() {
          *
          * @param param1 Parameter 1.
          * @param param2 Parameter 2.
-         * @return A new instance of fragment ButtonSeeTopics.
+         * @return A new instance of fragment likert.
          */
         // TODO: Rename and change types and number of parameters
         @JvmStatic
         fun newInstance(param1: String, param2: String) =
-            ButtonSeeTopics().apply {
+            likert().apply {
                 arguments = Bundle().apply {
                     putString(ARG_PARAM1, param1)
                     putString(ARG_PARAM2, param2)
