+++ app/src/main/java/com/educa/ui/fragments/FragmentModalDelete.kt
@@ -1,18 +1,14 @@
-package com.educa
+package com.educa.ui.fragments
 
 import android.os.Bundle
-import android.util.Log
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import android.widget.Button
-import android.widget.EditText
 import androidx.fragment.app.DialogFragment
+import com.educa.R
 import com.educa.api.model.Topic
 import com.educa.api.service.ApiClient
-import retrofit2.Call
-import retrofit2.Callback
-import retrofit2.Response
 
 
 class FragmentModalDelete : DialogFragment() {
