+++ app/src/main/res/layout/activity_access_thread.xml
@@ -12,7 +12,7 @@
 
         <fragment
             android:id="@+id/fg_bubble_question"
-            class="com.educa.FragmentEditContent"
+            class="com.educa.ui.fragments.FragmentEditContent"
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
             android:layout_marginTop="40dp"
@@ -138,7 +138,7 @@
 
     <fragment
         android:id="@+id/fg_menu"
-        class="com.educa.FragmentMenu"
+        class="com.educa.ui.fragments.FragmentMenu"
         android:layout_width="match_parent"
         android:layout_height="match_parent"
         tools:layout="@layout/fragment_menu" />
