+++ app/src/main/res/layout/activity_reading.xml
@@ -58,7 +58,7 @@
 
     <fragment
         android:id="@+id/fg_menu"
-        class="com.educa.FragmentMenu"
+        class="com.educa.ui.fragments.FragmentMenu"
         android:layout_width="match_parent"
         android:layout_height="match_parent"
         tools:layout="@layout/fragment_menu" />
