+++ app/src/main/res/layout/fragment_button_see_topics.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ButtonSeeTopics">
+    tools:context=".ui.fragments.ButtonSeeTopics">
 
     <RelativeLayout
         style="@style/style_btnBlueLarger"
