+++ app/src/main/res/layout/fragment_card_content.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="wrap_content"
-    tools:context=".FragmentCardContent">
+    tools:context=".ui.fragments.FragmentCardContent">
 
     <RelativeLayout
         android:id="@+id/card_view"
