+++ app/src/main/res/layout/fragment_card_topic.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="wrap_content"
-    tools:context=".FragmentEditContent">
+    tools:context=".ui.fragments.FragmentEditContent">
 
 
     <RelativeLayout
