+++ app/src/main/res/layout/fragment_menu.xml
@@ -4,7 +4,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:orientation="vertical"
-    tools:context=".FragmentMenu">
+    tools:context=".ui.fragments.FragmentMenu">
 
     <RelativeLayout
         android:id="@+id/drawerLayout"
