+++ app/src/main/res/layout/fragment_modal_update.xml
@@ -7,7 +7,7 @@
     android:layout_gravity="center_horizontal"
     android:background="#FF626262"
     android:orientation="vertical"
-    tools:context=".FragmentModalUpdate">
+    tools:context=".ui.fragments.FragmentModalUpdate">
 
     <LinearLayout
         android:id="@+id/linearLayout"
