+++ app/src/main/res/layout/fragment_navbar.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="wrap_content"
-    tools:context=".FragmentNavBar">
+    tools:context=".ui.fragments.FragmentNavBar">
 
     <RelativeLayout
         android:layout_width="match_parent"
