+++ app/src/main/res/layout/fragment_salutation.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="wrap_content"
-    tools:context=".FragmentSalutation">
+    tools:context=".ui.fragments.FragmentSalutation">
 
     <ImageView
         android:id="@+id/img_figure"
